Changes and Troubleshooting Front:
ResponsiveTable.vue has better design. Now each row has a edit button that leads to a offcanvas that lets you tp edit each product in the list.

EditProductModal.vue is the content in the offcanvas

FinalList.vue created in order to show the final test. You can specify amount and unit here and when you have already bought the product , just check the checkbox and it will be a double line on all of the row which indicates the it's off the list.
The Changes here are only html stuff and it's not making changes to lists in the backend.

ShoppingLists.vue has a load button that loads the list and it routes into a new page that is specified to each list. There you see SingleShoppingListView.vue. There you can edit or delete products from the loaded list

in index.js the new route for SingleShoppingListView.vue is specifed